requirements.txt
# ── Browser automation ──
selenium
undetected-chromedriver
webdriver-manager
setuptools               # distutils compat for Python 3.12+

# ── Audio CAPTCHA solving ──
SpeechRecognition
pydub
imageio-ffmpeg
PyAudio

# ── OpenAI (tour links, venue types) ──
openai

# ── API ──
fastapi
uvicorn[standard]
pydantic-settings
requests